Refactoring; Fixes; and other changes.
- Renamed quilting folder to jena2020 & parallel_quilting.py to quilting.py.
- Moved generateTextureMap from jena2020/generate.py to quilting.py; renamed it to generate_texture; replaced code sections with previously defined auxiliary methods.
- Fixed zero parallelization level to work with any version.
- Version 2 now converts image to Lab color space (supposes source is in RGB); this is only done to images, not to latent images. ***
- Changed find_patch_v3 behavior.

***
cv is now imported in nodes.py. Therefore, to use any node, cv must be installed.
note that in current implementation conversion to Lab format is done in the ImageQuilting node since it is only applicable when using images.
